Multi-session: REST sessions, one WebSocket per session, SessionHandle refactor
Server:
- Add GET /api/sessions REST endpoint with running boolean
- AgentRuntimeManager.sessionHasRuntime() to expose active runtimes

RookKit:
- Add RookAPI.sessions() REST method for session listing

Mac client:
- New SessionHandle class: owns dedicated AcpSocket + all per-session state
- ChatSessionController refactored to handle registry, loads sessions via REST
- Switching sessions preserves background session state

AS-BUILT-ARCHITECTURE: updated all docs for new REST + WebSocket-per-session shape

Issue #99
