+++ AS-BUILT-ARCHITECTURE/README.md
@@ -15,7 +15,7 @@ This directory replaces the old single-file architecture summary with per-surfac
 
 Rook is built around two protocol boundaries:
 
-1. client ↔ server: ACP over one connection-level WebSocket at `/api/ws`
+1. client ↔ server: session discovery over REST (`GET /api/sessions`); agent interaction over one ACP WebSocket per session at `/api/ws`
 2. server ↔ runtime: ACP over stdio subprocesses, one process per public session
 
 The server is the stable broker. Clients are thin native surfaces over the same REST + ACP contract.
