+++ server/src/server/index.ts
@@ -21,6 +21,7 @@ import { REPO_ROOT } from "./paths.js";
 import { registerEnvironmentRoutes } from "./routes/environmentRoutes.js";
 import { registerDiagnosticRoutes } from "./routes/diagnosticRoutes.js";
 import { registerRuntimeRoutes } from "./routes/runtimeRoutes.js";
+import { registerSessionRoutes } from "./routes/sessionRoutes.js";
 import { registerAcpFacadeRoute } from "./routes/acpFacadeRoute.js";
 import { ServerAuth } from "./auth.js";
 import { loadAgentRuntimes } from "./config/agentRuntimes.js";
@@ -112,6 +113,7 @@ export async function buildServer(options: BuildServerOptions = {}) {
 
   app.get("/api/health", async () => ({ ok: true, service: "rook" }));
   await registerRuntimeRoutes(app, runtimeManager);
+  await registerSessionRoutes(app, runtimeManager);
   await registerEnvironmentRoutes(app, environmentManager, environmentIdentifier, locationRegistrar, runtimeManager);
   await registerDiagnosticRoutes(app, environmentManager);
   await registerAcpFacadeRoute(app, runtimeManager, auth);
